Remove code duplicated from cats-kernel.
