--- laws/src/main/scala/algebra/laws/IsSerializable.scala
@@ -1,12 +1,0 @@
-package algebra.laws
-
-import scala.util.DynamicVariable
-
-/**
-  * Object with a dynamic variable that allows users to skip the
-  * serialization tests for certain instances.
-  */
-private[laws] object IsSerializable {
-  val runTests = new DynamicVariable[Boolean](true)
-  def apply(): Boolean = runTests.value
-}
