+++ laws/src/main/scala/algebra/laws/LatticePartialOrderLaws.scala
@@ -3,6 +3,8 @@ package algebra.laws
 import algebra._
 import algebra.lattice._
 
+import cats.kernel.laws._
+
 import org.typelevel.discipline.Laws
 
 import org.scalacheck.{Arbitrary, Prop}
