+++ laws/src/main/scala/algebra/laws/LogicLaws.scala
@@ -3,6 +3,8 @@ package algebra.laws
 import algebra._
 import algebra.lattice.{Heyting, Bool, GenBool}
 
+import cats.kernel.laws._
+
 import org.scalacheck.{Arbitrary, Prop}
 import org.scalacheck.Prop._
