+++ laws/src/main/scala/algebra/laws/RingLaws.scala
@@ -3,6 +3,8 @@ package laws
 
 import algebra.ring._
 
+import cats.kernel.laws._
+
 import org.typelevel.discipline.Predicate
 
 import org.scalacheck.{Arbitrary, Prop}
@@ -10,11 +12,16 @@ import org.scalacheck.Arbitrary._
 import org.scalacheck.Prop._
 
 object RingLaws {
-  def apply[A : Eq : Arbitrary](implicit pred0: Predicate[A]) = new RingLaws[A] {
+  def apply[A : Eq : Arbitrary: AdditiveMonoid]: RingLaws[A] =
+    withPred[A](new Predicate[A] {
+      def apply(a: A): Boolean = Eq[A].neqv(a, AdditiveMonoid[A].zero)
+    })
+
+  def withPred[A: Eq: Arbitrary](pred0: Predicate[A]): RingLaws[A] = new RingLaws[A] {
     def Arb = implicitly[Arbitrary[A]]
     def pred = pred0
     val nonZeroLaws = new GroupLaws[A] {
-      def Arb = Arbitrary(arbitrary[A] filter pred0)
+      def Arb = Arbitrary(arbitrary[A] filter pred)
       def Equ = Eq[A]
     }
   }
@@ -26,14 +33,49 @@ trait RingLaws[A] extends GroupLaws[A] {
   val nonZeroLaws: GroupLaws[A]
   def pred: Predicate[A]
 
-  def withPred(pred0: Predicate[A], replace: Boolean = true): RingLaws[A] = RingLaws[A](
-    Equ,
-    Arb,
-    if (replace) pred0 else pred && pred0
-  )
+  def withPred(pred0: Predicate[A], replace: Boolean = true): RingLaws[A] =
+    RingLaws.withPred(if (replace) pred0 else pred && pred0)(Equ, Arb)
 
   implicit def Arb: Arbitrary[A]
   implicit def Equ: Eq[A] = nonZeroLaws.Equ
+
+  // additive groups
+
+  def additiveSemigroup(implicit A: AdditiveSemigroup[A]) = new AdditiveProperties(
+    base = semigroup(A.additive),
+    parents = Nil,
+    Rules.serializable(A),
+    Rules.repeat1("sumN")(A.sumN),
+    Rules.repeat2("sumN", "+")(A.sumN)(A.plus)
+  )
+
+  def additiveCommutativeSemigroup(implicit A: AdditiveCommutativeSemigroup[A]) = new AdditiveProperties(
+    base = commutativeSemigroup(A.additive),
+    parents = List(additiveSemigroup)
+  )
+
+  def additiveMonoid(implicit A: AdditiveMonoid[A]) = new AdditiveProperties(
+    base = monoid(A.additive),
+    parents = List(additiveSemigroup),
+    Rules.repeat0("sumN", "zero", A.zero)(A.sumN),
+    Rules.collect0("sum", "zero", A.zero)(A.sum)
+  )
+
+  def additiveCommutativeMonoid(implicit A: AdditiveCommutativeMonoid[A]) = new AdditiveProperties(
+    base = commutativeMonoid(A.additive),
+    parents = List(additiveMonoid)
+  )
+
+  def additiveGroup(implicit A: AdditiveGroup[A]) = new AdditiveProperties(
+    base = group(A.additive),
+    parents = List(additiveMonoid),
+    Rules.consistentInverse("subtract")(A.minus)(A.plus)(A.negate)
+  )
+
+  def additiveCommutativeGroup(implicit A: AdditiveCommutativeGroup[A]) = new AdditiveProperties(
+    base = commutativeGroup(A.additive),
+    parents = List(additiveGroup)
+  )
 
   // multiplicative groups
 
@@ -176,6 +218,15 @@ trait RingLaws[A] extends GroupLaws[A] {
 
 
   // property classes
+
+  class AdditiveProperties(
+    val base: GroupProperties,
+    val parents: Seq[AdditiveProperties],
+    val props: (String, Prop)*
+  ) extends RuleSet {
+    val name = base.name
+    val bases = List("base" -> base)
+  }
 
   class MultiplicativeProperties(
     val base: GroupLaws[A] => GroupLaws[A]#GroupProperties,
